Exception Handling

+++ src/main/java/com/lopeztw/ltcommerce/services/ProductService.java
@@ -1,14 +1,18 @@
 package com.lopeztw.ltcommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.lopeztw.ltcommerce.dto.ProductDTO;
 import com.lopeztw.ltcommerce.entities.Product;
 import com.lopeztw.ltcommerce.repositories.ProductRepository;
+import com.lopeztw.ltcommerce.services.exceptions.DataBaseException;
+import com.lopeztw.ltcommerce.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class ProductService {
@@ -20,7 +24,7 @@ public class ProductService {
 	
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Product product = repository.findById(id).get();
+		Product product = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado!")); // orElseThore tenta acessar um obj e caso n encontre, lança uma excessao
 		return new ProductDTO(product);
 		
 	}
@@ -50,9 +54,17 @@ public class ProductService {
 		return new ProductDTO(entity);
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+	        	repository.deleteById(id);    		
+		}
+	    	catch (DataIntegrityViolationException e) { // caso tente deletar algo que esteja atrelado a outra coisa, ex: produto e pedido
+	        	throw new DataBaseException("Falha de integridade referencial");
+	   	}
 	}
 
 	private void copyDtoToEntity(ProductDTO dto, Product entity) {
